ISSUE #1499 - moved addcomment logic from issue to model/comment file
